requirements.txt
python-dotenv
openai
ipykernel
ipywidgets
matplotlib
#azure-ai-vision==0.15.1b1
azure-ai-vision
diffusers
transformers
scipy
ftfy
accelerate
azure-ai-ml
azure-identity
mlflow
azureml-mlflow
tenacity
tqdm
faiss-cpu
bokeh
torch
torchvision
azure-functions
requests
pyodbc
azure-communication-networktraversal
azure-communication-identity
